organize types and move credits form component
